+++ Options/CHANGELOG.lua
@@ -9,6 +9,7 @@ TMW.CHANGELOG = [==[
 ### Bug Fixes
 * Classic: Fixed the Unit Class condition's options.
 * Classic: Fixed the Weapon Imbue icon type & Condition for offhands.
+* Classic: Fixed combopoint tracking.
 
 ## v8.6.9
 * Classic: Aura durations now account for learned talents.